Componetize E2EFilterTest
Extracts format-agnostic parts of E2EFilterTest into base class
E2EFilterTestBase. Provides a reusable test framework for adding
selective readers for other file formats.
